requirements.txt
chainlit
elasticsearch==8.15.1
langchain==0.3.4
langchain_core==0.3.13
langchain_elasticsearch==0.3.0
langchain_experimental==0.3.2
langchain_openai==0.2.3
numpy==2.1.2
openai_whisper==20240930
pdf2image==1.17.0
Pillow==11.0.0
pytesseract==0.3.13
python-dotenv==1.0.1
rich==13.9.3
sentence_transformers==3.2.0
transformers==4.45.2